fawa_requirements.txt
absl-py==0.10.0
advertorch==0.2.3
argon2-cffi==20.1.0
astor==0.8.1
attrs==20.1.0
backcall==0.2.0
bleach==3.1.5
certifi==2020.6.20
cffi==1.14.2
chardet==3.0.4
cycler==0.10.0
decorator==4.4.2
defusedxml==0.6.0
dominate==2.5.2
entrypoints==0.3
fastprogress==0.1.20
future==0.18.2
gast==0.4.0
grpcio==1.32.0
h5py==2.10.0
idna==2.10
imageio==2.9.0
importlib-metadata==1.7.0
ipykernel==5.3.4
ipython==7.18.1
ipython-genutils==0.2.0
ipywidgets==7.5.1
jedi==0.17.2
Jinja2==2.11.2
joblib==0.16.0
jsonpatch==1.26
jsonpointer==2.0
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.7
jupyter-console==6.2.0
jupyter-core==4.6.3
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.2.0
Markdown==3.2.2
MarkupSafe==1.1.1
matplotlib==3.3.1
mistune==0.8.4
mkl-fft==1.0.12
mkl-random==1.1.1
mock==4.0.2
nbconvert==5.6.1
nbformat==5.0.7
networkx==2.5
notebook==6.1.3
numpy==1.16.4
olefile==0.46
opencv-python==4.4.0.42
packaging==20.4
pandocfilters==1.4.2
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow @ file:///tmp/build/80754af9/pillow_1594307325547/work
prometheus-client==0.8.0
prompt-toolkit==3.0.7
protobuf==3.13.0
ptyprocess==0.6.0
pycparser==2.20
Pygments==2.6.1
pyparsing==2.4.7
pyrsistent==0.16.0
python-dateutil==2.8.1
pytorch-msssim==0.2.1
PyWavelets==1.1.1
pyzmq==19.0.1
qtconsole==4.7.7
QtPy==1.9.0
requests==2.24.0
scikit-image==0.17.2
scikit-learn==0.23.2
scikit-opt==0.5.7
scipy==1.5.2
Send2Trash==1.5.0
six==1.15.0
tensorboard==1.13.1
tensorflow==1.13.1
tensorflow-estimator==1.13.0
termcolor==1.1.0
terminado==0.8.3
testpath==0.4.4
threadpoolctl==2.1.0
tifffile==2020.8.13
torch==1.6.0
torchfile==0.1.0
torchgan==0.0.4
torchvision==0.7.0
tornado==6.0.4
tqdm==4.50.0
traitlets==5.0.3
urllib3==1.25.10
visdom==0.1.8.9
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==0.57.0
Werkzeug==1.0.1
widgetsnbextension==3.5.1
zipp==3.1.0
